Move more traits to primitives.
