+++ pallets/common/src/traits/group.rs
@@ -13,8 +13,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use crate::traits::CommonConfig;
-
 use polymesh_primitives::{traits::group::MemberCount, IdentityId};
 
 use frame_support::{
@@ -34,7 +32,12 @@ pub trait WeightInfo {
     fn abdicate_membership() -> Weight;
 }
 
-pub trait Config<I>: CommonConfig + pallet_timestamp::Config + pallet_identity::Config {
+pub trait Config<I>:
+    frame_system::Config
+    + pallet_permissions::Config
+    + pallet_timestamp::Config
+    + pallet_identity::Config
+{
     /// The overarching event type.
     type RuntimeEvent: From<Event<Self, I>> + Into<<Self as frame_system::Config>::RuntimeEvent>;
