+++ pallets/common/src/traits/portfolio.rs
@@ -17,88 +17,15 @@
 //!
 //! The interface allows to accept portfolio custody
 
-use crate::{nft::NFTTrait, CommonConfig};
 use frame_support::decl_event;
-use frame_support::dispatch::DispatchResult;
 use frame_support::pallet_prelude::Get;
 use frame_support::weights::Weight;
 use polymesh_primitives::asset::AssetId;
-use polymesh_primitives::traits::AssetFnConfig;
+use polymesh_primitives::traits::{AssetFnConfig, NFTTrait};
 use polymesh_primitives::{
-    Balance, Fund, FundDescription, IdentityId, Memo, NFTId, PortfolioId, PortfolioName,
-    PortfolioNumber, SecondaryKey,
+    Fund, FundDescription, IdentityId, Memo, PortfolioId, PortfolioName, PortfolioNumber,
 };
 use sp_std::vec::Vec;
-
-/// This trait is used to accept custody of a portfolio
-pub trait PortfolioSubTrait<AccountId> {
-    /// Checks that the custodian is authorized for the portfolio
-    ///
-    /// # Arguments
-    /// * `portfolio` - Portfolio to check
-    /// * `custodian` - DID of the custodian
-    fn ensure_portfolio_custody(portfolio: PortfolioId, custodian: IdentityId) -> DispatchResult;
-
-    /// Ensure that the `portfolio` exists.
-    ///
-    /// # Arguments
-    /// * `portfolio` - Portfolio to check
-    fn ensure_portfolio_validity(portfolio: &PortfolioId) -> DispatchResult;
-
-    /// Locks some tokens of a portfolio
-    ///
-    /// # Arguments
-    /// * `portfolio` - Portfolio to lock tokens
-    /// * `asset_id` - [`AssetId`] of the token to lock
-    /// * `amount` - Amount of tokens to lock
-
-    fn lock_tokens(portfolio: &PortfolioId, asset_id: &AssetId, amount: Balance) -> DispatchResult;
-
-    /// Unlocks some tokens of a portfolio
-    ///
-    /// # Arguments
-    /// * `portfolio` - Portfolio to unlock tokens
-    /// * asset_id` - [`AssetId`] of the token to unlock
-    /// * `amount` - Amount of tokens to unlock
-    fn unlock_tokens(
-        portfolio: &PortfolioId,
-        asset_id: &AssetId,
-        amount: Balance,
-    ) -> DispatchResult;
-
-    /// Ensures that the portfolio's custody is with the provided identity
-    /// And the secondary key has the relevant portfolio permission
-    ///
-    /// # Arguments
-    /// * `portfolio` - PortfolioId of the portfolio to check
-    /// * `custodian` - Identity of the custodian
-    /// * `secondary_key` - Secondary key that is accessing the portfolio
-    fn ensure_portfolio_custody_and_permission(
-        portfolio: PortfolioId,
-        custodian: IdentityId,
-        secondary_key: Option<&SecondaryKey<AccountId>>,
-    ) -> DispatchResult;
-
-    /// Locks the given nft. This prevents transfering the same NFT more than once.
-    ///
-    /// # Arguments
-    /// * `portfolio_id` - PortfolioId that contains the nft to be locked.
-    /// asset_id` - [`AssetId`] of the NFT.
-    /// * `nft_id` - the id of the nft to be unlocked.
-    fn lock_nft(portfolio_id: &PortfolioId, asset_id: &AssetId, nft_id: &NFTId) -> DispatchResult;
-
-    /// Unlocks the given nft.
-    ///
-    /// # Arguments
-    /// * `portfolio_id` - PortfolioId that contains the locked nft.
-    /// asset_id` - [`AssetId`] of the NFT.
-    /// * `nft_id` - the id of the nft to be unlocked.
-    fn unlock_nft(portfolio_id: &PortfolioId, asset_id: &AssetId, nft_id: &NFTId)
-        -> DispatchResult;
-
-    /// Returns `true` if the portfolio has pre-approved the receivement of `asset_id`, otherwise returns `false`.
-    fn skip_portfolio_affirmation(portfolio_id: &PortfolioId, asset_id: &AssetId) -> bool;
-}
 
 pub trait WeightInfo {
     fn create_portfolio(l: u32) -> Weight;
@@ -118,7 +45,9 @@ pub trait WeightInfo {
     fn create_custody_portfolio() -> Weight;
 }
 
-pub trait Config: CommonConfig + pallet_identity::Config + AssetFnConfig {
+pub trait Config:
+    frame_system::Config + pallet_permissions::Config + pallet_identity::Config + AssetFnConfig
+{
     type RuntimeEvent: From<Event> + Into<<Self as frame_system::Config>::RuntimeEvent>;
     type WeightInfo: WeightInfo;
     /// Maximum number of fungible assets that can be moved in a single transfer call.
