+++ pallets/runtime/tests/src/staking/mock.rs
@@ -42,7 +42,6 @@ use sp_staking::{EraIndex, SessionIndex};
 use pallet_staking::types::SlashingSwitch;
 use pallet_staking::{self as pallet_staking, *};
 use polymesh_common_utilities::traits::balances::{AccountData, CheckCdd};
-use polymesh_common_utilities::traits::portfolio::PortfolioSubTrait;
 use polymesh_common_utilities::traits::CommonConfig;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::constants::currency::POLY;
@@ -50,7 +49,7 @@ use polymesh_primitives::identity_id::GenesisIdentityRecord;
 use polymesh_primitives::{
     traits::{
         group::{GroupTrait, InactiveMember},
-        SubsidiserTrait,
+        PortfolioSubTrait, SubsidiserTrait,
     },
     Authorization, AuthorizationData, Claim, IdentityId, Moment, NFTId, Permissions, PortfolioId,
     SecondaryKey, Signatory,
